Add AdbServerDadbTest
