--- nohup.out
@@ -1,25 +1,0 @@
-WARNING | unexpected system image feature string, emulator might not function correctly, please try updating the emulator.
-WARNING | cannot add library /Users/leland/.android-sdk/emulator/qemu/darwin-aarch64/lib64/vulkan/libvulkan.dylib: failed
-INFO    | Warning: QMetaObject::connectSlotsByName: No matching signal for on_new_posture_requested(int) ((null):0, (null))
-
-INFO    | Warning: QMetaObject::connectSlotsByName: No matching signal for on_dismiss_posture_selection_dialog() ((null):0, (null))
-
-INFO    | Warning: QMetaObject::connectSlotsByName: No matching signal for on_new_resizable_requested(PresetEmulatorSizeType) ((null):0, (null))
-
-INFO    | Warning: QMetaObject::connectSlotsByName: No matching signal for on_dismiss_resizable_dialog() ((null):0, (null))
-
-INFO    | configAndStartRenderer: setting vsync to 60 hz
-INFO    | added library /Users/leland/.android-sdk/emulator/lib64/vulkan/libvulkan.dylib
-INFO    | Started GRPC server at 127.0.0.1:8554, security: Local
-INFO    | Advertising in: /Users/leland/Library/Caches/TemporaryItems/avd/running/pid_29641.ini
-INFO    | Your emulator is out of date, please update by launching Android Studio:
- - Start Android Studio
- - Select menu "Tools > Android > SDK Manager"
- - Click "SDK Tools" tab
- - Check "Android Emulator" checkbox
- - Click "OK"
-
-INFO    | boot completed
-INFO    | boot time 18642 ms
-INFO    | Increasing screen off timeout, logcat buffer size to 2M.
-INFO    | Revoking microphone permissions for Google App.
